requirements.txt
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.7.0
cachetools==5.3.2
certifi @ file:///croot/certifi_1690232220950/work/certifi
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dill==0.3.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gradio==4.1.1
gradio_client==0.7.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.18.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mdurl==0.1.2
numpy==1.20.3
openpyxl==3.1.2
orjson==3.9.10
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
protobuf==4.25.0
pyarrow==14.0.1
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pydub==0.25.1
pyFTS==1.6
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
streamlit==1.28.2
tenacity==8.2.3
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
websockets==11.0.3
zipp==3.17.0
